refactor: Movie 객체의 식별자를 Long으로 변경 후 Adapter의 item id 함수의 반환값을 movie 아이템의 id로 변경

+++ app/src/main/java/woowacourse/movie/MovieListActivity.kt
@@ -17,7 +17,7 @@ class MovieListActivity : AppCompatActivity(), MovieListContract.View {
         movies.adapter = presenter.getAdapter()
     }
 
-    override fun navigateToTicketing(movieId: Int) {
+    override fun navigateToTicketing(movieId: Long) {
         Intent(this, TicketingActivity::class.java).apply {
             putExtra(EXTRA_MOVIE_ID, movieId)
             startActivity(this)
